Moved info page into its own component
