+++ src/App.js
@@ -1,105 +1,21 @@
 import React, { useState, useRef, useEffect } from 'react';
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+  Link
+} from "react-router-dom";
 
-import Step from './components/displays/step';
-import NavBar from './components/displays/sideNav';
-import Hidden from '@material-ui/core/Hidden';
-
-import { stepIDName } from './css.js'
-
-const Steps = require("./data/config");
+import Info from './pages/info';
 
 export default function App() {
-  const [currentStepNumber, setCurrentStepNumber] = useState(1); // Hook for visible section state
-  const headerRef = useRef(null); // Header height reference
-
-  const doScroll = id => {
-    document.getElementById(id).scrollIntoView({
-      behavior: "smooth",
-      block: "start"
-    });
-  };
-
-  // Gets the top and bottom y offsets of a given element
-  const getDimensions = ele => {
-    const { height } = ele.getBoundingClientRect();
-    const offsetTop = ele.offsetTop;
-    const offsetBottom = offsetTop + height;
-
-    return {
-      height,
-      offsetTop,
-      offsetBottom
-    };
-  };
-
-  // Finds the current visible section
-  useEffect(() => {
-    const handleScroll = () => {
-      const { height: headerHeight } = getDimensions(headerRef.current);
-      const scrollPosition = window.scrollY + headerHeight;
-
-      const currentIndex = Steps.steps.findIndex( (_, index) =>  {
-        const ele = document.getElementById(stepIDName(index + 1))
-
-        if (ele) {
-          const { offsetBottom, offsetTop } = getDimensions(ele);
-          return scrollPosition >= offsetTop && scrollPosition < offsetBottom;
-        }
-        return false
-      });
-
-      if (currentIndex === undefined) {
-        return 
-      };
-
-      const stepNumber = currentIndex + 1
-
-      if (stepNumber !== currentStepNumber) {
-        setCurrentStepNumber(stepNumber);
-      }
-    };
-
-    window.addEventListener("scroll", handleScroll);
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, [currentStepNumber]);
 
   return (
     <div className='app'>
-      <div className='header' ref={headerRef}>
+      <div className='header' >
         <h1>Clear My Missouri Record</h1>
       </div>
-
-      <div className='content'>
-        <div className='steps'>
-          {Steps.steps.map((data, index) => {
-            return (
-              <Step
-                key={index.toString()}
-                stepNumber={index + 1}
-                {...data}
-              />
-            );
-          })}
-        </div>
-      </div>
-      <Hidden mdDown>
-        <div className='navbar'>
-          <ol className='navbar-sticky'>
-            {Steps.steps.map((step, index) => {
-              return (
-                <NavBar
-                  key={index.toString()}
-                  onClick={() => doScroll(stepIDName(index + 1))}
-                  isCurrent={currentStepNumber === index + 1}
-                  name={step.navName}
-                />
-              );
-            })}
-          </ol>
-        </div>
-      </Hidden>
+      <Info></Info>
       <div className='bottom-spacer' />
       <div className='footer'>
         <div className='footer-content'>footer placeholder</div>
